+++ examples/cloth-alembic/main.cpp
@@ -1,8 +1,8 @@
+#include <elasty/alembic-manager.hpp>
 #include <elasty/cloth-sim-object.hpp>
 #include <elasty/constraint.hpp>
 #include <elasty/engine.hpp>
 #include <elasty/particle.hpp>
-#include <elasty/utils.hpp>
 #include <timer.hpp>
 
 // #define SPHERE_COLLISION
