+++ examples/pbd-xpbd-comparison/main.cpp
@@ -1,3 +1,4 @@
+#include <elasty/alembic-manager.hpp>
 #include <elasty/cloth-sim-object.hpp>
 #include <elasty/constraint.hpp>
 #include <elasty/engine.hpp>
