+++ include/elasty/utils.hpp
@@ -12,14 +12,6 @@ namespace elasty
     class AbstractConstraint;
     class ClothSimObject;
 
-    /// \brief An abstract class for making alembic export easier.
-    class AlembicManager
-    {
-    public:
-        /// \brief Submit the current frame to the managed alembic file.
-        virtual void submitCurrentStatus() = 0;
-    };
-
     constexpr double pi() { return 3.14159265358979323846264338327950288; }
 
     /// \param particles an array of N particles
@@ -39,11 +31,6 @@ namespace elasty
                                        const Eigen::Vector3d&                            fixed_position,
                                        const std::vector<std::shared_ptr<Particle>>&     particles,
                                        std::vector<std::shared_ptr<AbstractConstraint>>& constraints);
-
-    /// \brief A utility function to instantiating an alembic manager tailered for cloth objects
-    std::shared_ptr<AlembicManager> createClothAlembicManager(const std::string&                    output_file_path,
-                                                              const std::shared_ptr<ClothSimObject> cloth_sim_object,
-                                                              const double                          delta_time);
 
     /// \brief Export the current cloth state as an obj file.
     void exportCurrentClothStateAsObj(const std::string&                    output_file_path,
